requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
requests
djangorestframework
